Swift: Rename SymbolClass.Symbol to Component
This is to avoid ambiguity when a DFA is used as a partition (where strings are elements in the partition, not symbols). Other good names might be Member or Item. Element would be best, but is taken up by Swift for too many things. String is another keyword I'd like to use, but is taken up by Swift.

Also rename PartNo to Label, to be consistent with PartitionedDictionaryProtocol.